Move ghutil to under shared
